added license competency and fix entities

+++ app/Services/Domain/ProgramService.php
@@ -62,12 +62,11 @@ class ProgramService
      * Create new StudyProgram
      *
      * @param Collection $data
-     * @param array $licenses
      * @param bool $org
      * @param bool $flush
      * @return StudyProgram
      */
-    public function create(Collection $data, array $licenses = [], $org = false, $flush = true)
+    public function create(Collection $data, $org = false, $flush = true)
     {
         $program = new StudyProgram;
         $program->setIdDikti($data->get('id_dikti'));
@@ -90,9 +89,9 @@ class ProgramService
         if ($org instanceof Organization) {
             $program->setOrg($org);
         }
-        
-        if($licenses){
-            $this->setLicenses($program, $licenses);
+
+        if($data->get('license', [])){
+            $this->setLicenses($program, $data->get('license'));
         }
 
         EntityManager::persist($program);
@@ -109,12 +108,11 @@ class ProgramService
      *
      * @param StudyProgram $program
      * @param Collection $data
-     * @param array $licenses
      * @param bool $org
      * @param bool $flush
      * @return StudyProgram
      */
-    public function update(StudyProgram $program, Collection $data, array $licenses = [], $org = false, $flush = true)
+    public function update(StudyProgram $program, Collection $data, $org = false, $flush = true)
     {
         $program->setIdDikti($data->get('id_dikti'));
         $program->setCode($data->get('code'));
@@ -137,8 +135,8 @@ class ProgramService
             $program->setOrg($org);
         }
 
-        if($licenses){
-            $this->setLicenses($program, $licenses);
+        if($data->get('license', [])){
+            $this->setLicenses($program, $data->get('license'), 'update');
         }
 
         EntityManager::persist($program);
@@ -151,19 +149,22 @@ class ProgramService
     }
 
     /**
-     * Update license program
+     * Set license program
      *
      * @param StudyProgram $studyProgram
      * @param array $licenses
+     * @param string $type
      */
-    private function setLicenses(StudyProgram $studyProgram, array $licenses = [])
+    private function setLicenses(StudyProgram $studyProgram, array $licenses = [], $type = 'create')
     {
         /** @var LicenseService $licenseService */
         $licenseService = app(LicenseService::class);
         /** @var LicenseProgramService $licenseProgramService */
         $licenseProgramService = app(LicenseProgramService::class);
 
-        $licenseProgramService->delete($studyProgram);
+        if ($type == 'update') {
+            $licenseProgramService->delete($studyProgram);
+        }
 
         if (count($licenses)) {
             foreach ($licenses as $licenseId) {
